+++ packages/apollo-server-express/src/ApolloServer.test.ts
@@ -2,6 +2,9 @@ import { expect } from 'chai';
 import 'mocha';
 import express from 'express';
 
+import net from 'net';
+import http from 'http';
+
 import request from 'request';
 import FormData from 'form-data';
 import fs from 'fs';
@@ -9,7 +12,12 @@ import fetch from 'node-fetch';
 import { createApolloFetch } from 'apollo-fetch';
 
 import { ApolloServerBase, AuthenticationError } from 'apollo-server-core';
-import { registerServer } from './ApolloServer';
+import { registerServer, ApolloServer } from './ApolloServer';
+
+import {
+  testApolloServer,
+  createServerInfo,
+} from 'apollo-server-integration-testsuite';
 
 //to remove the circular dependency, we reference it directly
 const gql = require('../../apollo-server/dist/index').gql;
@@ -26,22 +34,43 @@ const resolvers = {
   },
 };
 
+const url = 'http://localhost:4000/graphql';
+const uri = url;
+
+describe('apollo-server-express', () => {
+  let server;
+  let httpServer;
+  testApolloServer(
+    async options => {
+      server = new ApolloServer(options);
+      const app = express();
+      registerServer({ app, server });
+      httpServer = await new Promise<http.Server>(resolve => {
+        const s = app.listen({ port: 4000 }, () => resolve(s));
+      });
+      return createServerInfo(server, httpServer);
+    },
+    async () => {
+      if (server) await server.stop();
+      if (httpServer && httpServer.listening) await httpServer.close();
+    },
+  );
+});
+
 describe('apollo-server-express', () => {
   //to remove the circular dependency, we reference it directly
   const ApolloServer = require('../../apollo-server/dist/index').ApolloServer;
-  let server: ApolloServerBase & {
-    createGraphQLServerOptions: (
-      req: express.Request,
-      res: express.Response,
-    ) => any;
-  };
+  let server: ApolloServerBase | any;
+
   let app: express.Application;
+  let httpServer: http.Server;
 
   afterEach(async () => {
     if (server) await server.stop();
+    if (httpServer) await httpServer.close();
   });
 
-  describe('', () => {
+  describe('constructor', () => {
     it('accepts typeDefs and resolvers', () => {
       const app = express();
       const server = new ApolloServer({ typeDefs, resolvers });
@@ -65,7 +94,9 @@ describe('apollo-server-express', () => {
 
       registerServer({ app, server });
 
-      const { url: uri } = await server.listen();
+      httpServer = await new Promise<http.Server>(resolve => {
+        const l = app.listen({ port: 4000 }, () => resolve(l));
+      });
       const apolloFetch = createApolloFetch({ uri });
       const result = await apolloFetch({ query: '{hello}' });
 
@@ -96,7 +127,9 @@ describe('apollo-server-express', () => {
 
       registerServer({ app, server, gui: true });
 
-      const { url } = await server.listen();
+      httpServer = await new Promise<http.Server>(resolve => {
+        const l = app.listen({ port: 4000 }, () => resolve(l));
+      });
       const apolloFetch = createApolloFetch({ uri: url });
       const result = await apolloFetch({ query: INTROSPECTION_QUERY });
 
@@ -105,7 +138,7 @@ describe('apollo-server-express', () => {
         'GRAPHQL_VALIDATION_FAILED',
       );
 
-      return new Promise((resolve, reject) => {
+      return new Promise<http.Server>((resolve, reject) => {
         request(
           {
             url,
@@ -140,8 +173,10 @@ describe('apollo-server-express', () => {
 
       registerServer({ app, server });
 
-      const { url } = await server.listen();
-      return new Promise((resolve, reject) => {
+      httpServer = await new Promise<http.Server>(resolve => {
+        const l = app.listen({ port: 4000 }, () => resolve(l));
+      });
+      return new Promise<http.Server>((resolve, reject) => {
         request(
           {
             url,
@@ -174,7 +209,9 @@ describe('apollo-server-express', () => {
 
       registerServer({ app, server, cors: { origin: 'apollographql.com' } });
 
-      const { url: uri } = await server.listen({});
+      httpServer = await new Promise<http.Server>(resolve => {
+        const l = app.listen({ port: 4000 }, () => resolve(l));
+      });
 
       const apolloFetch = createApolloFetch({ uri }).useAfter(
         (response, next) => {
@@ -196,7 +233,9 @@ describe('apollo-server-express', () => {
 
       registerServer({ app, server, bodyParserConfig: { limit: 0 } });
 
-      const { url: uri } = await server.listen({});
+      httpServer = await new Promise<http.Server>(resolve => {
+        const l = app.listen({ port: 4000 }, () => resolve(l));
+      });
 
       const apolloFetch = createApolloFetch({ uri });
 
@@ -226,7 +265,10 @@ describe('apollo-server-express', () => {
 
         registerServer({ app, server, bodyParserConfig: { limit: 0 } });
 
-        const { port } = await server.listen();
+        httpServer = await new Promise<http.Server>(resolve => {
+          const l = app.listen({ port: 4000 }, () => resolve(l));
+        });
+        const { port } = httpServer.address() as net.AddressInfo;
 
         return new Promise((resolve, reject) => {
           request(
@@ -262,7 +304,10 @@ describe('apollo-server-express', () => {
           },
         });
 
-        const { port } = await server.listen({});
+        httpServer = await new Promise<http.Server>(resolve => {
+          const l = app.listen({ port: 4000 }, () => resolve(l));
+        });
+        const { port } = httpServer.address() as net.AddressInfo;
 
         return new Promise((resolve, reject) => {
           request(
@@ -295,7 +340,10 @@ describe('apollo-server-express', () => {
           disableHealthCheck: true,
         });
 
-        const { port } = await server.listen({});
+        httpServer = await new Promise<http.Server>(resolve => {
+          const l = app.listen({ port: 4000 }, () => resolve(l));
+        });
+        const { port } = httpServer.address() as net.AddressInfo;
 
         return new Promise((resolve, reject) => {
           request(
@@ -351,7 +399,10 @@ describe('apollo-server-express', () => {
           server,
         });
 
-        const { port } = await server.listen({});
+        httpServer = await new Promise<http.Server>(resolve => {
+          const l = app.listen({ port: 4000 }, () => resolve(l));
+        });
+        const { port } = httpServer.address() as net.AddressInfo;
 
         const body = new FormData();
 
@@ -423,7 +474,9 @@ describe('apollo-server-express', () => {
         app = express();
         registerServer({ app, server });
 
-        const { url: uri } = await server.listen();
+        httpServer = await new Promise<http.Server>(resolve => {
+          const l = app.listen({ port: 4000 }, () => resolve(l));
+        });
         const apolloFetch = createApolloFetch({ uri });
 
         const result = await apolloFetch({ query: '{hello}' });
@@ -461,7 +514,9 @@ describe('apollo-server-express', () => {
         app = express();
         registerServer({ app, server });
 
-        const { url: uri } = await server.listen();
+        httpServer = await new Promise<http.Server>(resolve => {
+          const l = app.listen({ port: 4000 }, () => resolve(l));
+        });
         const apolloFetch = createApolloFetch({ uri });
 
         const result = await apolloFetch({ query: `{error}` });
@@ -499,7 +554,9 @@ describe('apollo-server-express', () => {
         app = express();
         registerServer({ app, server });
 
-        const { url: uri } = await server.listen();
+        httpServer = await new Promise<http.Server>(resolve => {
+          const l = app.listen({ port: 4000 }, () => resolve(l));
+        });
         const apolloFetch = createApolloFetch({ uri });
 
         const result = await apolloFetch({ query: `{error}` });
@@ -536,7 +593,9 @@ describe('apollo-server-express', () => {
         app = express();
         registerServer({ app, server });
 
-        const { url: uri } = await server.listen();
+        httpServer = await new Promise<http.Server>(resolve => {
+          const l = app.listen({ port: 4000 }, () => resolve(l));
+        });
         const apolloFetch = createApolloFetch({ uri });
 
         const result = await apolloFetch({ query: `{error}` });
