+++ packages/apollo-server-integration-testsuite/package.json
@@ -31,8 +31,14 @@
     "definition": "dist/index.d.ts"
   },
   "devDependencies": {
-    "apollo-link-persisted-queries": "^0.2.0",
+    "apollo-fetch": "^0.7.0",
+    "apollo-link": "^1.2.2",
+    "apollo-link-http": "^1.5.4",
+    "apollo-link-persisted-queries": "^0.2.1",
+    "graphql-subscriptions": "^0.5.8",
     "graphql-tag": "^2.9.2",
-    "js-sha256": "^0.9.0"
+    "js-sha256": "^0.9.0",
+    "subscriptions-transport-ws": "^0.9.11",
+    "ws": "^5.2.0"
   }
 }
