+++ test/tests.js
@@ -12,22 +12,20 @@ process.on('unhandledRejection', reason => {
 require('../packages/apollo-server-core/dist/runQuery.test.js');
 require('../packages/apollo-server-core/dist/runHttpQuery.test.js');
 require('../packages/apollo-server-core/dist/errors.test.js');
-require('../packages/apollo-server-core/dist/ApolloServer.test.js');
 
 //Apollo server 2 tests
 
 //apollo-server
 require('../packages/apollo-server/dist/index.test.js');
 
-//apollo-server-express
-require('../packages/apollo-server-express/dist/ApolloServer.test.js');
-
-//Apollo server 1 tests
 require('../packages/apollo-server-module-operation-store/dist/operationStore.test');
 
+require('../packages/apollo-server-express/dist/ApolloServer.test.js');
 require('../packages/apollo-server-express/dist/expressApollo.test');
 require('../packages/apollo-server-express/dist/connectApollo.test');
+
 (NODE_MAJOR_VERSION >= 9 ||
   (NODE_MAJOR_VERSION >= 8 && NODE_MAJOR_REVISION >= 9)) &&
-  require('../packages/apollo-server-hapi/dist/hapiApollo.test'); // Hapi 17 is 8.9+
+require('../packages/apollo-server-hapi/dist/hapiApollo.test') && // Hapi 17 is 8.9+
+  require('../packages/apollo-server-hapi/dist/ApolloServer.test.js');
 require('../packages/apollo-server-express/dist/apolloServerHttp.test');
